refactor the code and modularize it
